Started FlashDevelop project. Created Card, Deck and Suit classes.
The Suit class is just an enumeration of the five different suits: four
minor arcana and the major arcana.

The Card class contains information about the card's suit and number. It
can return a string for the name of the card, such as "Ace of Cups".

The deck generates a vector of cards.